Fix tests & reorganize tests
