+++ Tests/AnalyticsTests/ComScore/ComScoreTrackerTests.swift
@@ -111,54 +111,6 @@ final class ComScoreTrackerTests: ComScoreTestCase {
         }
     }
 
-    func testSeekWhilePlaying() {
-        let player = Player(item: .simple(
-            url: Stream.onDemand.url,
-            metadata: AssetMetadataMock(),
-            trackerAdapters: [
-                ComScoreTracker.adapter { _ in ["meta": "data"] }
-            ]
-        ))
-
-        player.play()
-        expect(player.playbackState).toEventually(equal(.playing))
-
-        expectAtLeastEvents(
-            .pause { labels in
-                expect(labels.ns_st_po).to(beCloseTo(0, within: 0.5))
-            },
-            .play { labels in
-                expect(labels.ns_st_po).to(beCloseTo(7, within: 0.5))
-            }
-        ) {
-            player.seek(at(.init(value: 7, timescale: 1)))
-        }
-    }
-
-    func testSeekWhilePause() {
-        let player = Player(item: .simple(
-            url: Stream.onDemand.url,
-            metadata: AssetMetadataMock(),
-            trackerAdapters: [
-                ComScoreTracker.adapter { _ in ["meta": "data"] }
-            ]
-        ))
-
-        expect(player.playbackState).toEventually(equal(.paused))
-
-        expectNoEvents(during: .seconds(2)) {
-            player.seek(at(.init(value: 7, timescale: 1)))
-        }
-
-        expectAtLeastEvents(
-            .play { labels in
-                expect(labels.ns_st_po).to(beCloseTo(7, within: 0.5))
-            }
-        ) {
-            player.play()
-        }
-    }
-
     func testFailure() {
         let player = Player(item: .simple(
             url: Stream.unavailable.url,
@@ -169,113 +121,6 @@ final class ComScoreTrackerTests: ComScoreTestCase {
         ))
 
         expectNoEvents(during: .seconds(3)) {
-            player.play()
-        }
-    }
-
-    func testOnDemandDvrProperties() {
-        let player = Player(item: .simple(
-            url: Stream.onDemand.url,
-            metadata: AssetMetadataMock(),
-            trackerAdapters: [
-                ComScoreTracker.adapter { _ in ["meta": "data"] }
-            ]
-        ))
-
-        expectAtLeastEvents(
-            .play { labels in
-                expect(labels.ns_st_ldw).to(equal(0))
-                expect(labels.ns_st_ldo).to(equal(0))
-            }
-        ) {
-            player.play()
-        }
-    }
-
-    func testLiveDvrProperties() {
-        let player = Player(item: .simple(
-            url: Stream.live.url,
-            metadata: AssetMetadataMock(),
-            trackerAdapters: [
-                ComScoreTracker.adapter { _ in ["meta": "data"] }
-            ]
-        ))
-
-        expectAtLeastEvents(
-            .play { labels in
-                expect(labels.ns_st_ldw).to(equal(0))
-                expect(labels.ns_st_ldo).to(equal(0))
-            }
-        ) {
-            player.play()
-        }
-    }
-
-    func testDvrPropertiesAtLiveEdge() {
-        let player = Player(item: .simple(
-            url: Stream.dvr.url,
-            metadata: AssetMetadataMock(),
-            trackerAdapters: [
-                ComScoreTracker.adapter { _ in ["meta": "data"] }
-            ]
-        ))
-
-        expectAtLeastEvents(
-            .play { labels in
-                expect(labels.ns_st_ldo).to(equal(0))
-                expect(labels.ns_st_ldw).to(equal(Stream.dvr.duration.seconds))
-            }
-        ) {
-            player.play()
-        }
-    }
-
-    func testDvrPropertiesAwayFromLiveEdge() {
-        let player = Player(item: .simple(
-            url: Stream.dvr.url,
-            metadata: AssetMetadataMock(),
-            trackerAdapters: [
-                ComScoreTracker.adapter { _ in ["meta": "data"] }
-            ]
-        ))
-
-        player.play()
-        expect(player.playbackState).toEventually(equal(.playing))
-
-        expectAtLeastEvents(
-            .pause { labels in
-                expect(labels.ns_st_ldo).to(equal(0))
-                expect(labels.ns_st_ldw).to(equal(Stream.dvr.duration.seconds))
-            },
-            .play { labels in
-                expect(labels.ns_st_ldo).to(equal(4))
-                expect(labels.ns_st_ldw).to(equal(Stream.dvr.duration.seconds))
-            }
-        ) {
-            player.seek(at(player.timeRange.end - CMTime(value: 4, timescale: 1)))
-        }
-    }
-
-    func testMetadata() {
-        let player = Player(item: .simple(
-            url: Stream.dvr.url,
-            metadata: AssetMetadataMock(),
-            trackerAdapters: [
-                ComScoreTracker.adapter { _ in
-                    [
-                        "meta_1": "custom-1",
-                        "meta_2": "42"
-                    ]
-                }
-            ]
-        ))
-
-        expectAtLeastEvents(
-            .play { labels in
-                expect(labels.metadata("meta_1")).to(equal("custom-1"))
-                expect(labels.metadata("meta_2")).to(equal(42))
-            }
-        ) {
             player.play()
         }
     }
